+++ workspaces/intelligent-assistant/plugins/intelligent-assistant/src/components/Attachment.tsx
@@ -15,11 +15,9 @@
  */
 
 import { makeStyles } from '@material-ui/core';
-import {
-  AttachmentEdit,
-  ChatbotDisplayMode,
-  PreviewAttachment,
-} from '@patternfly/chatbot';
+import { ChatbotDisplayMode } from '@patternfly/chatbot';
+import { AttachmentEdit } from '@patternfly/chatbot/dist/dynamic/AttachmentEdit';
+import { PreviewAttachment } from '@patternfly/chatbot/dist/dynamic/PreviewAttachment';
 
 import { useTranslation } from '../hooks/useTranslation';
 import { useFileAttachmentContext } from './AttachmentContext';
@@ -84,7 +82,7 @@ const Attachment = () => {
         secondaryActionButtonText={t('modal.cancel')}
         primaryActionButtonText={t('modal.save')}
         modalFooterClassName={classes.modalFooter}
-        onSave={(_, content) => {
+        onSave={(_event, content) => {
           setCurrentFileContent({
             ...currentFileContent,
             content,