+++ workspaces/boost/plugins/boost-connector-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @red-hat-developer-hub/backstage-plugin-boost-connector-utils
 
+## 0.3.2
+
+### Patch Changes
+
+- a65f815: bump workspace to Backstage 1.54.6
+
+## 0.3.1
+
+### Patch Changes
+
+- 0e772a4: bump workspace to Backstage 1.54.4
+
 ## 0.3.0
 
 ### Minor Changes